Refactering for routing
