+++ app/controllers/users/passwords_controller.rb
@@ -1,0 +1,2 @@
+class Users::PasswordsController < Devise::PasswordsController
+end
