+++ app/controllers/users/sessions_controller.rb
@@ -1,0 +1,2 @@
+class Users::SessionsController < Devise::SessionsController
+end
