+++ app/policies/user_policy.rb
@@ -1,5 +1,21 @@
 class UserPolicy < ApplicationPolicy
-  def display?
+  def index?
+    user.has_role? :admin or user.has_role? :staff
+  end
+
+  def show?
+    user.has_role? :admin or user.has_role? :staff
+  end
+
+  def create?
+    user.has_role? :admin or user.has_role? :staff
+  end
+
+  def destroy?
+    user.has_role? :admin or user.has_role? :staff or user.id == record.id
+  end
+
+  def show_image?
     user.has_role? :admin or user.has_role? :staff
   end
 
@@ -8,6 +24,10 @@ class UserPolicy < ApplicationPolicy
   end
 
   def checkout?
+    user.has_role? :admin or user.has_role? :staff
+  end
+
+  def checkout_all?
     user.has_role? :admin or user.has_role? :staff
   end
 end
