+++ app/views/displays/show.html.erb
