+++ app/views/layouts/_display_ad.html.erb
@@ -1,4 +1,4 @@
-<section class="displayprof bg_black" style="background-image: url(<%= adimg_path(ad.id) %>);">
+<section class="displayprof bg_black" style="background-image: url(<%= advertisement_image_path(ad.id) %>);">
     <div class="sponsor_info_area">
       <h1 class="sponsor">Powered by <%= ad.sponsor %></h1>
       <% if ad.url.present? %>
