+++ app/views/layouts/_display_user.html.erb
@@ -3,7 +3,7 @@
     <div class="post_profimg">
       <div class="mainimgframe">
         <% if user.profimg %>
-          <%= image_tag show_profimg_user_path(user.id) %>
+          <%= image_tag user_image_path(user.id) %>
         <% end %>
       </div>
     </div>
