+++ config/routes.rb
@@ -1,69 +1,62 @@
 Rails.application.routes.draw do
   # Root
-  root to: 'app#index'
+  root to: "app#index"
 
-  #Term of Service
-  get 'tos', to: 'app#tos'
+  # Static Pages
+  get "tos", to: "app#tos"
+  get "privacy", to: "app#privacy"
 
-  #Privacy Policy
-  get 'privacy', to: 'app#privacy'
+  # Devise
+  devise_for :users,
+    path: "",
+    path_names: {
+      sign_in: "login",
+      sign_out: "logout",
+      registration: "signup",
+      sign_up: "",
+    },
+    controllers: {
+      sessions: "users/sessions",
+      passwords: "users/passwords",
+      registrations: "users/registrations",
+      confirmations: "users/confirmations",
+      omniauth_callbacks: "users/omniauth_callbacks",
+    }
 
-  # Profile
-  get 'users/:user_id', to: 'users#show', as: 'profile'
+  # User
+  resources :users, only: [:show, :destroy] do
+    get "image", to: "users#show_image"
+  end
+  get "user/destroy_confirmation", to: "users#destroy_confirmation"
+
+  # Advertisement Image
+  resources :advertisements, only: [:index, :create] do
+    get "image", to: "advertisements#show_image"
+  end
 
   # Display
-  get 'display/:shop_id', to: 'users#display', as: 'display'
-
-  # Review
-  resources :reviews, only: [:create]
+  resources :displays, only: [:show]
 
   # Contact
   resource :contact, only: [:show, :create]
 
-  # Check In/Out
-  get 'checkin/:user_id', to: 'users#checkin', as: 'checkin'
-  get 'checkout/:user_id', to: 'users#checkout', as: 'checkout'
-  get 'checkout_all', to: 'users#checkout_all', as: 'checkout_all'
+  # Review
+  resource :review, only: [:create]
 
-  #Destroy
-  get 'destroy_confirmation', to: 'users#destroy_confirmation', as: 'destroy_confirmation'
-  get 'destroy', to: 'users#destroy', as: 'destroy'
+  # Checkin
+  get "checkin/:id", to: "checkins#checkin", as: "checkin"
+  get "checkout/:id", to: "checkins#checkout", as: "checkout"
+  get "checkout_all", to: "checkins#checkout_all", as: "checkout_all"
 
-  # Devise
-  devise_for :users,
-    path: '',
-    path_names: {
-      sign_in: 'login',
-      sign_out: 'logout',
-      registration: 'signup',
-      sign_up: '',
-    },
-    controllers: {
-      registrations: 'users/registrations',
-      confirmations: 'users/confirmations',
-      omniauth_callbacks: 'users/omniauth_callbacks',
-    }
-
-  # Profile Image
-  resources :users do
-    member do
-      get 'show_profimg'
-    end
-  end
-
-  # Advertisement Image
-  resource :advertisement, only: [:new, :create]
-  get 'ads/:ad_id/image', to: 'advertisements#show_adimg', as: 'adimg'
-
-  # Admin
+  # Admin Pages
   ActiveAdmin.routes(self)
 
   # Web Mailer
   if Rails.env.development?
-    mount LetterOpenerWeb::Engine, at: '/letter_opener'
+    mount LetterOpenerWeb::Engine, at: "/letter_opener"
   end
 
   # Exception
-  get '*not_found', to: 'application#routing_error'
-  post '*not_found', to: 'application#routing_error'
+  get "*not_found", to: "application#routing_error"
+  post "*not_found", to: "application#routing_error"
 end
